+++ test/scrapers/tvmaze/testTvMazeHelper.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "scrapers/tv_show/tvmaze/TvMazeApi.h"
+
+mediaelch::scraper::TvMazeApi& getTvMazeApi();