+++ test/unit/tv_shows/testTvMazeId.cpp
@@ -18,8 +18,11 @@ TEST_CASE("TvMazeId data type", "[data]")
     {
         CHECK_FALSE(TvMazeId().isValid());
         CHECK_FALSE(TvMazeId("").isValid());
+        CHECK_FALSE(TvMazeId(0).isValid());
+        CHECK_FALSE(TvMazeId(-1).isValid());
 
         CHECK(TvMazeId("262504").isValid());
+        CHECK(TvMazeId(262504).isValid());
     }
 
     SECTION("Conversion") { CHECK(TvMazeId("262504").toString() == "262504"); }