+++ CHANGELOG.md
@@ -15,6 +15,10 @@
 
  - Concert: Add IMDb/TMDb ID fields
  - Added `*.webm` to default file filters for video files (#1122)
+ - Add TVmaze as a TV scraper (#834)
+   TVmaze works great for actor images and season posters.  Due to API
+   limitations not all episode details are available that you can see
+   on their website.
 
 ### Internal Improvements and Changes